+++ layouts/default.vue
@@ -1,31 +1,26 @@
 <template>
-    <div class="slidev-layout">
-        <!-- 顶部 Banner -->
+    <div class="slidev-layout default">
         <div class="flex items-center gap-5 mb-8">
             <img :src="logoUrl" alt="Lab banner" class="h-16 w-auto rounded-lg" />
             <div class="flex flex-col gap-1 text-left">
-                <!-- 从 slot 的第一个 h1 里取出来 -->
                 <div class="text-3xl">{{ titleText }}</div>
                 <div class="text-xs text-[#5b6b82] tracking-wide">{{ topic }}</div>
             </div>
         </div>
 
-        <!-- 过滤掉第一个 h1 的 slot 内容 -->
-        <div>
+        <div class="my-auto">
             <component v-for="(node, i) in filteredNodes" :is="node" :key="i" />
         </div>
-
-        <Toc />
     </div>
 </template>
 
 <script setup lang="ts">
 import { useSlideContext } from '@slidev/client'
-import { computed, h, useSlots } from 'vue'
+import { ref, onMounted, useSlots } from 'vue'
 
 const { $slidev } = useSlideContext();
 const configs = $slidev.configs as any;
-const logoUrl = configs.logoUrl || '';
+const logoUrl = configs.logoUrl || 'https://picx.zhimg.com/80/v2-32c67121f7b3b7395aec3ad20fb97713_1440w.png';
 const topic = configs.topic || '';
 
 // 获取默认 slot 的内容