+++ CHANGELOG.md
@@ -196,7 +196,7 @@
 * add component TextWithOptionalLink ([d0ddd3f](https://github.com/alexanderdavide/slidev-theme-academic/commit/d0ddd3f195919a4a44fc79c6dd6022fd78304945))
 * add components Footnotes and Footnote ([cef139b](https://github.com/alexanderdavide/slidev-theme-academic/commit/cef139b35e71973acc003b2bc2bc213e3c9462ae))
 * add layout figure ([47822da](https://github.com/alexanderdavide/slidev-theme-academic/commit/47822dadc9b37dbfd9c0d64c44472922a632d0b2))
-* add layout figure-side ([6a98eb1](https://github.com/alexanderdavide/slidev-theme-academic/commit/6a98eb13497f6dd2ad9b2aab471782f789abba6e))
+* add layout image-text ([6a98eb1](https://github.com/alexanderdavide/slidev-theme-academic/commit/6a98eb13497f6dd2ad9b2aab471782f789abba6e))
 * add layout table-of-contents ([e428743](https://github.com/alexanderdavide/slidev-theme-academic/commit/e4287433a60743ce49e11a7bff3ff7eb60dcb741))
 * add/overwrite layout cover ([630edbd](https://github.com/alexanderdavide/slidev-theme-academic/commit/630edbd01434cb07bb04ac56adc83c2e3f9129d5))
 * define max height and center ([80dad54](https://github.com/alexanderdavide/slidev-theme-academic/commit/80dad543d86797f65fbbd1b0fa12ead00c29c727))